+++ src/axom/quest/detail/clipping/MonotonicZSORClipper.hpp
@@ -1,0 +1,222 @@
+// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
+// other Axom Project Developers. See the top-level COPYRIGHT file for details.
+//
+// SPDX-License-Identifier: (BSD-3-Clause)
+
+#ifndef AXOM_QUEST_MONOTONICZSORCLIPPER_HPP
+#define AXOM_QUEST_MONOTONICZSORCLIPPER_HPP
+
+#include "axom/klee/Geometry.hpp"
+#include "axom/quest/MeshClipperStrategy.hpp"
+#include "axom/primal/geometry/CoordinateTransformer.hpp"
+
+namespace axom
+{
+namespace quest
+{
+namespace experimental
+{
+
+/*!
+ * @brief Geometry clipping operations for simple 3D
+ * surface-of-revolution geometries.
+ *
+ * This implementation requires the SOR curve r(z) to be monotonic in z.
+ * The values of z that make up the polyline specifying the curve must
+ * be either non-decreasing or non-increasing.  A curve may have segments
+ * with vertical slope, but it may not double back in z.
+ * For SOR curves that double back, use \a SORClipper.
+ *
+ * The SOR specification may include axis orientation and location
+ * in addition to any external transformation.
+ */
+class MonotonicZSORClipper : public MeshClipperStrategy
+{
+public:
+  /*!
+   * @brief Constructor.
+   *
+   * @param [in] kGeom Describes the shape to place
+   *   into the mesh.
+   * @param [in] name To override the default strategy name
+   *
+   * \c kGeom.asHierarchy() must contain the following data:
+   * - sorOrigin: 3D coordinates of the point (z,r) = (0,0)
+   * - sorDirection: a vector in the direction of the SOR axis.
+   * - discreteFunction: The discretized r(z) curve as an array of (z,r) pairs.
+   * - levelOfRefinement: number of refinement levels used
+   *   to approximate the sphere with octahedra.  The number
+   *   of octs grows exponentially with level
+   *   (@see quest::discretize(const axom::ArrayView<Point2D> &polyline, int pointcount, int levels, axom::Array<OctType> &out, int &octcount)).
+   *   In practice, keep this number to 11 or less.
+   */
+  MonotonicZSORClipper(const klee::Geometry& kGeom, const std::string& name = "");
+
+  /*!
+   * @brief Construct with parameters to override the specifications
+   * in the klee::Geometry.
+   */
+  MonotonicZSORClipper(const klee::Geometry& kGeom,
+                       const std::string& name,
+                       axom::ArrayView<const Point2DType> discreteFunction,
+                       const Point3DType& sorOrigin,
+                       const Vector3DType& sorDirection,
+                       axom::IndexType levelOfRefinement);
+
+  virtual ~MonotonicZSORClipper() = default;
+
+  const std::string& name() const override { return m_name; }
+
+  bool labelCellsInOut(quest::experimental::ShapeMesh& shappeMesh,
+                       axom::Array<LabelType>& label) override;
+
+  bool labelTetsInOut(quest::experimental::ShapeMesh& shapeMesh,
+                      axom::ArrayView<const axom::IndexType> cellIds,
+                      axom::Array<LabelType>& tetLabels) override;
+
+  bool getGeometryAsOcts(quest::experimental::ShapeMesh& shappeMesh,
+                         axom::Array<OctahedronType>& octs) override;
+
+  axom::ArrayView<const Point2DType> getSorCurve() const { return m_sorCurve.view(); }
+
+  //@{
+  //! @name Utilities shared with SORClipper for handling SOR.
+  /*!
+   * @brief Find division points between curve sections where z (x)
+   * changes directions.
+   *
+   * @param sorCurve Set of at least 2 2D points describing a curve
+   *   in r-z space (in host array).
+   *
+   * @return Indices of switchbacks, plus the first and last indices.
+   */
+  static axom::Array<axom::IndexType> findZSwitchbacks(axom::ArrayView<const Point2DType> pts);
+
+  /*
+   * @brief Combine consecutive radial segments of the curve into a
+   * single segment.
+   *
+   * This step is necessary because some other steps assume there are
+   * no consecutive radial segments.
+   */
+  static void combineRadialSegments(axom::Array<Point2DType>& sorCurve);
+  //@}
+
+#if !defined(__CUDACC__)
+private:
+#endif
+  std::string m_name;
+
+  /*!
+   * @brief The discrete r(z) curve as an array of y(x) points.
+   *
+   * This data is before internal or external transformations.
+   * It may include points on each end to connect the curve to
+   * the axis of rotation.
+   */
+  axom::Array<Point2DType> m_sorCurve;
+
+  //! @brief Bounding box of points in m_sorCurve;
+  BoundingBox2DType m_curveBb;
+
+  //! @brief Maximum radius of the SOR.
+  double m_maxRadius;
+
+  //! @brief Minimum radius of the SOR.
+  double m_minRadius;
+
+  //!@brief The point corresponding to z=0 on the SOR axis.
+  Point3DType m_sorOrigin;
+
+  //!@brief SOR axis in 3D space, in the direction of increasing z.
+  Vector3DType m_sorDirection;
+
+  //!@brief Level of refinement for discretizing curved
+  // analytical shapes and surfaces of revolutions.
+  axom::IndexType m_levelOfRefinement = 0;
+
+  /*!
+   * @brief Boxes (in rz space) on the curve.
+   *
+   * The curve lies completely in these boxes and includes the planes
+   * of the base and top.  Points in these boxes are require more
+   * computation to determine their signed distance.
+   */
+  axom::Array<BoundingBox2DType> m_bbOn;
+
+  /*!
+   * @brief Boxes (in rz space) completely under the curve.
+   *
+   * These boxes lie completely under the curve.
+   */
+  axom::Array<BoundingBox2DType> m_bbUnder;
+
+  //!@brief Internal and external transforms (includes m_sorDirection and m_sorOrigin).
+  axom::primal::experimental::CoordinateTransformer<double> m_transformer;
+
+  /*!
+   * @brief Inverse of m_transformer.
+   *
+   * Axom supports vector scaling.  @see axom::klee::Scale.  This means
+   * a SOR may be transformed into a shape that we cannot represent.
+   * Therefore, we don't transform the shape until after it's discretized.
+   * When needed, we will inverse-transform the mesh.
+   */
+  axom::primal::experimental::CoordinateTransformer<double> m_invTransformer;
+
+  template <typename ExecSpace>
+  void labelCellsInOutImpl(quest::experimental::ShapeMesh& shapeMesh,
+                           axom::ArrayView<LabelType> label);
+
+  template <typename ExecSpace>
+  void labelTetsInOutImpl(quest::experimental::ShapeMesh& shapeMesh,
+                          axom::ArrayView<const axom::IndexType> cellIds,
+                          axom::ArrayView<LabelType> tetLabels);
+
+  template <typename ExecSpace>
+  void computeCurveBoxes(quest::experimental::ShapeMesh& shapeMesh,
+                         axom::Array<BoundingBox2DType>& bbOn,
+                         axom::Array<BoundingBox2DType>& bbUnder);
+
+  /*!
+   * @brief Compute 2D bounding box of a polyhedron in the r-z plane.
+   * @tparam PolyhedronType Either TetrahedronType or HexahedronType.
+   */
+  template <typename PolyhedronType>
+  AXOM_HOST_DEVICE BoundingBox2DType estimateBoundingBoxInRz(const PolyhedronType& vertices);
+
+  AXOM_HOST_DEVICE inline MeshClipperStrategy::LabelType rzBbToLabel(
+    const BoundingBox2DType& bbInRz,
+    const axom::ArrayView<const BoundingBox2DType>& bbOn,
+    const axom::ArrayView<const BoundingBox2DType>& bbUnder);
+
+  // Extract clipper info from MeshClipperStrategy::m_info.
+  void extractClipperInfo();
+
+  /*!
+   * @brief Subdivide large segments of the SOR curve to make
+   * screening more precise.
+   *
+   * @param sorCurveIn [in] Un-divided SOR curve
+   * @param maxMean [in] Subdivide segment if the harmonic mean
+   *   of its bounding box sides exceeds this value.
+   * @param maxDz [in] Subdivide segment if its length along the
+   *   axis of symmetry exceeds this value.
+   * @param minDz [in] Don't subdivide segments below this dz.
+   */
+  axom::Array<Point2DType> subdivideCurve(const Array<Point2DType>& sorCurveIn,
+                                          double maxMean,
+                                          double maxDz,
+                                          double minDz);
+
+  //!@brief Compute geometry as octs, by policy.
+  template <typename ExecSpace>
+  bool getGeometryAsOctsImpl(quest::experimental::ShapeMesh& shappeMesh,
+                             axom::Array<OctahedronType>& octs);
+};
+
+}  // namespace experimental
+}  // namespace quest
+}  // namespace axom
+
+#endif  // AXOM_QUEST_FSORCLIPPER_HPP