+++ RELEASE-NOTES.md
@@ -48,6 +48,9 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Quest: Adds a 3D winding number example and impoves WN performance in 2D and 3D
 
 ###  Changed
+- Version of `quest::discretize` that approximates a surface-of-revolution from a polyline
+  now respects the allocator ID of the output `Array`.  It no longer resets the ID to the
+  execution space default.
 - Evaluation methods for line integrals in `axom::primal` have been generalized, and 
   `evaluate_scalar_line_integral` has been renamed to `evaluate_line_integral`.
 - Treatment of materials on strided-structured Blueprint meshes has changed in `axom::mir`.